requirements.txt
numpy
h5py
PyYAML
fire
pytest